requirements.txt
# Python 3.12+
Django==5.1.2
djangorestframework==3.15.2
django-filter==24.3
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
psycopg2-binary==2.9.9  # For future PostgreSQL migration

# Image processing
Pillow==10.4.0

# API documentation
drf-spectacular==0.27.2

# Development tools
django-debug-toolbar==4.4.6
ipython==8.27.0

# Testing
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1

# Telegram Bot
python-telegram-bot==21.6
httpx==0.27.2  # For async requests

